Put Clan Subcommands in own classes
