+++ src/main/java/xyz/gamlin/clans/commands/SubCommands/Clan/ClanSetHomeSubCommand.java
@@ -1,0 +1,40 @@
+package xyz.gamlin.clans.commands.SubCommands.Clan;
+
+import org.bukkit.command.CommandSender;
+import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.entity.Player;
+import xyz.gamlin.clans.Clans;
+import xyz.gamlin.clans.models.Clan;
+import xyz.gamlin.clans.utils.ClansStorageUtil;
+import xyz.gamlin.clans.utils.ColorUtils;
+
+public class ClanSetHomeSubCommand {
+    FileConfiguration clansConfig = Clans.getPlugin().getConfig();
+    FileConfiguration messagesConfig = Clans.getPlugin().messagesFileManager.getMessagesConfig();
+    public boolean setClanHomeSubCommand(CommandSender sender) {
+        if (sender instanceof Player) {
+            Player player = ((Player) sender).getPlayer();
+            if (clansConfig.getBoolean("clan-home.enabled")){
+                if (ClansStorageUtil.isClanOwner(player)){
+                    if (ClansStorageUtil.findClanByOwner(player) != null){
+                        Clan clan = ClansStorageUtil.findClanByOwner(player);
+                        clan.setClanHomeWorld(player.getLocation().getWorld().getName());
+                        clan.setClanHomeX(player.getLocation().getX());
+                        clan.setClanHomeY(player.getLocation().getY());
+                        clan.setClanHomeZ(player.getLocation().getZ());
+                        clan.setClanHomeYaw(player.getLocation().getYaw());
+                        clan.setClanHomePitch(player.getLocation().getPitch());
+                        player.sendMessage(ColorUtils.translateColorCodes(messagesConfig.getString("successfully-set-clan-home")));
+                    }
+                }else {
+                    player.sendMessage(ColorUtils.translateColorCodes(messagesConfig.getString("clan-must-be-owner")));
+                }
+            }else {
+                player.sendMessage(ColorUtils.translateColorCodes(messagesConfig.getString("function-disabled")));
+            }
+            return true;
+
+        }
+        return false;
+    }
+}
